transform_svg.sh - transform SVG to PS map XML
A small tool to convert a properly formatted SVG file to an XML file
suitable for importing in the PlaneShift Map Editor. XML files have a size
of 800x600 pixels and can be resized as required in the game.

To use, enter a command like the following:
    ./transform_svg.sh drawing.svg > drawing.xml
The drawing.xml file can be placed in PS's sketches directory and imported.